infra/scripts/requirements.txt
azure-identity>=1.19.0
python-dotenv==1.0.1
azure-cosmos>=4.7.0
rich